pong-game/backend/requirements.txt
# WS related
Django==4.2.7
daphne==4.0.0
channels_redis==4.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
python-dotenv==1.0.0
# Other dependencies
asgiref==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.2
wait-for-it==2.3.0
gunicorn==21.2.0
better_profanity==0.7.0
django-crontab==0.7.1
pillow==11.0.0
uuid==1.30